tools/image_pipeline/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
Pillow
pillow-heif
opencv-python
numpy
